pyproject.toml
[project]
name = "deepthroath"
version = "0.1.0"
description = "Universal RAG Evaluation and Red Teaming Framework"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deepteam>=1.0.6",
    "deepeval==3.9.7",
    "anthropic>=0.40.0",
    "openai>=1.50.0,<2.0.0",
    "pandas==2.3.3",
    "pyarrow==23.0.1",
    "pydantic==2.12.5",
    "fastapi==0.115.6",
    "uvicorn[standard]==0.34.0",
    "pyyaml>=6.0",
    "python-dotenv==1.2.2",
    "jinja2>=3.1.0",
    "weasyprint>=62.0",
    "gigachat",
    "httpx>=0.28.0",
    "croniter>=2.0.0",
    "ragas>=0.2.0",
    "langchain-openai>=0.1.0",
    "langchain-core>=0.2.0",
]

[tool.uv]
managed = true

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "I", "B"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "ruff>=0.15.11",
]
